horsemanshop/ui/serializers.py
from django.contrib.auth.models import User
from rest_framework import serializers

from horsemanshop.shop.models import Article


class UsersSerializer(serializers.ModelSerializer):

    articles = serializers.PrimaryKeyRelatedField(many=True, queryset=Article.objects.all())

    class Meta:
        model = User
        fields = ['id', 'username', 'articles']


class UserSerializer(serializers.ModelSerializer):
    user = serializers.PrimaryKeyRelatedField(read_only=True, default=serializers.CurrentUserDefault())
    articles = serializers.PrimaryKeyRelatedField(many=True, queryset=Article.objects.all())

    class Meta:
        model = User
        fields = ['id', 'username', 'articles', 'user']
